Implement look command.
The player may now use `/` to look around the visible map area and will see
what is in each location in the status area.

* Edit tile names to fit into sentences.

* Add `map_bounds` help function in `UI` class to replace specific code in
`draw_area` such that it can be used any time the visible portion of the map
needs to be determined.

* `look` method in `UI` class now requires a world object and uses it to
determine what is visible at each location.

* Add a `description_at` method to `World` class which returns the most
visible description of a location.