Fix creature None bug.
When we moved the map generation out of the World constructor we
changed the order in which things happen. The constructor would
populate the world with creatures, then we would build a map, and
sometimes the roads of that map would overwrite creatures in the
tile-space. When they attempted to move, we would have a None creature
attempting to move and crash.

Fixes #24