Improve UI and turn interation.
When the player selects invalid actions on his turn, time would
advance despite the hero not getting a turn. Now, when an actor
spends zero action points they get to act again immediatly. This can
be used for chaining actions or when an invalid action is selected
and the actor should choose again.

Fix #44
Fix #45