Fix map rendering near northwest corner.
Previously we changed left and top to be non-zero when the map was
smaller than the area in the x or y direction. Since we render in a
range from top to bottom and left to right, we would render those
parts incorrectly. We now leave the boundery negative but just skip
negative coordinates when redering.

Fix #50.